+++ sdl3-main/src/lib.rs
@@ -10,6 +10,8 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
+#[cfg(all(feature = "nightly", feature = "log-errors"))]
+use core::fmt::Display;
 #[cfg(feature = "nightly")]
 use core::{convert::Infallible, ops::FromResidual};
 
@@ -462,7 +464,20 @@ impl From<SDL_AppResult> for AppResult {
     }
 }
 
-#[cfg(feature = "nightly")]
+#[cfg(all(feature = "nightly", feature = "log-errors"))]
+impl<E: ::core::fmt::Display> FromResidual<Result<Infallible, E>> for AppResult {
+    fn from_residual(residual: Result<Infallible, E>) -> Self {
+        let Err(err) = residual;
+        let err = ::alloc::format!("{err}\0");
+        unsafe {
+            ::sdl3_sys::log::SDL_LogError(0, c"%s".as_ptr(), err.as_ptr());
+            ::sdl3_sys::error::SDL_SetError(c"%s".as_ptr(), err.as_ptr());
+        };
+        AppResult::Failure
+    }
+}
+
+#[cfg(all(feature = "nightly", not(feature = "log-errors")))]
 impl<E> FromResidual<Result<Infallible, E>> for AppResult {
     #[inline(always)]
     fn from_residual(_residual: Result<Infallible, E>) -> Self {
@@ -491,7 +506,22 @@ pub enum AppResultWithState<S: AppState> {
     Failure(Option<S>),
 }
 
-#[cfg(feature = "nightly")]
+#[cfg(all(feature = "nightly", feature = "log-errors"))]
+impl<S: AppState, E: ::core::fmt::Display> FromResidual<Result<Infallible, E>>
+    for AppResultWithState<S>
+{
+    fn from_residual(residual: Result<Infallible, E>) -> Self {
+        let Err(err) = residual;
+        let err = ::alloc::format!("{err}\0");
+        unsafe {
+            ::sdl3_sys::log::SDL_LogError(0, c"%s".as_ptr(), err.as_ptr());
+            ::sdl3_sys::error::SDL_SetError(c"%s".as_ptr(), err.as_ptr());
+        };
+        AppResultWithState::Failure(None)
+    }
+}
+
+#[cfg(all(feature = "nightly", not(feature = "log-errors")))]
 impl<S: AppState, E> FromResidual<Result<Infallible, E>> for AppResultWithState<S> {
     #[inline(always)]
     fn from_residual(_residual: Result<Infallible, E>) -> Self {