+++ sdl3-main/ChangeLog.md
@@ -6,6 +6,7 @@
       It still exists for compatibility, but it's hidden in the docs.
     - Pass arguments to main on `std` (`!std` already did)
     - Add `MainThreadData::assert_new/get/get_mut`
+    - Log error when converting to `AppResult*`
 
 - 0.5.4:
     - Fix accidental dependency on Rust 1.84